Corrections as discussed in the pyiron meeting
