+++ pyiron_base/__init__.py
@@ -12,7 +12,8 @@ from pyiron_base.storage.datacontainer import DataContainer
 from pyiron_base.storage.inputlist import InputList
 from pyiron_base.storage.parameters import GenericParameters
 from pyiron_base.storage.filedata import load_file, FileDataTemplate, FileData
-from pyiron_base.utils.util import Deprecator, deprecate, deprecate_soon, ImportAlarm
+from pyiron_base.utils.deprecate import Deprecator, deprecate, deprecate_soon
+from pyiron_base.utils.error import ImportAlarm
 from pyiron_base.jobs.job.extension.executable import Executable
 from pyiron_base.project.external import Notebook, load, dump
 from pyiron_base.jobs.job.generic import GenericJob
@@ -38,7 +39,7 @@ from pyiron_base.jobs.job.extension.server.queuestatus import validate_que_reque
 from pyiron_base.state.settings import Settings
 from pyiron_base.state.install import install_dialog
 from pyiron_base.jobs.datamining import PyironTable, TableJob
-from pyiron_base.utils.object import HasDatabase, HasStorage, PyironObject
+from pyiron_base.interfaces.object import HasDatabase, HasStorage, PyironObject
 from pyiron_base.database.performance import get_database_statistics
 from pyiron_base.interfaces.has_groups import HasGroups
 from pyiron_base.interfaces.has_hdf import HasHDF
