+++ pyiron_base/database/filetable.py
@@ -4,7 +4,7 @@ import pandas
 import datetime
 import h5io
 from pyfileindex import PyFileIndex
-from pyiron_base.utils.util import Singleton
+from pyiron_base.interfaces.singleton import Singleton
 from pyiron_base.database.generic import IsDatabase
