+++ pyiron_base/database/manager.py
@@ -7,7 +7,7 @@ A class for mediating connections to SQL databases.
 
 from urllib.parse import quote_plus
 from pyiron_base.state.logger import logger
-from pyiron_base.utils.util import Singleton
+from pyiron_base.interfaces.singleton import Singleton
 from pyiron_base.state.settings import settings as s
 from pyiron_base.database.generic import DatabaseAccess
 import os
