+++ pyiron_base/interfaces/singleton.py
@@ -1,0 +1,36 @@
+# coding: utf-8
+# Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
+# Distributed under the terms of "New BSD License", see the LICENSE file.
+
+"""
+Utility functions used in pyiron.
+In order to be accessible from anywhere in pyiron, they *must* remain free of any imports from pyiron!
+"""
+from abc import ABCMeta
+
+__author__ = "Joerg Neugebauer, Jan Janssen"
+__copyright__ = (
+    "Copyright 2020, Max-Planck-Institut für Eisenforschung GmbH - "
+    "Computational Materials Design (CM) Department"
+)
+__version__ = "1.0"
+__maintainer__ = "Jan Janssen"
+__email__ = "janssen@mpie.de"
+__status__ = "production"
+__date__ = "Sep 1, 2017"
+
+
+class Singleton(ABCMeta):
+    """
+    Implemented with suggestions from
+
+    http://stackoverflow.com/questions/6760685/creating-a-singleton-in-python
+
+    """
+
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
