+++ pyiron_base/jobs/job/template.py
@@ -6,7 +6,7 @@ Template class to define jobs
 """
 
 from pyiron_base.jobs.job.generic import GenericJob
-from pyiron_base.utils.object import HasStorage
+from pyiron_base.interfaces.object import HasStorage
 from pyiron_base.jobs.job.jobtype import JobType
 
 __author__ = "Jan Janssen"
