+++ pyiron_base/state/__init__.py
@@ -19,7 +19,7 @@ from pyiron_base.database.manager import database as _database
 from pyiron_base.state.publications import publications as _publications
 from pyiron_base.state.queue_adapter import queue_adapters as _queue_adapters
 from pyiron_base.state.settings import settings as _settings
-from pyiron_base.utils.util import Singleton
+from pyiron_base.interfaces.singleton import Singleton
 from typing import Dict, Union
 
 __author__ = "Liam Huber"
