+++ pyiron_base/storage/filedata.py
@@ -11,7 +11,7 @@ from functools import lru_cache
 import pandas
 
 from pyiron_base.storage.hdfio import FileHDFio, ProjectHDFio
-from pyiron_base.utils.util import ImportAlarm
+from pyiron_base.utils.error import ImportAlarm
 
 __author__ = "Niklas Siemer"
 __copyright__ = (
