+++ pyiron_base/storage/inputlist.py
@@ -3,7 +3,7 @@ Backwards compatible way of importing the DataContainer.
 """
 
 from pyiron_base.storage.datacontainer import DataContainer
-from pyiron_base.utils.util import deprecate
+from pyiron_base.utils.deprecate import deprecate
 
 
 class InputList(DataContainer):
