+++ pyiron_base/utils/deprecate.py
@@ -6,7 +6,6 @@
 Utility functions used in pyiron.
 In order to be accessible from anywhere in pyiron, they *must* remain free of any imports from pyiron!
 """
-from abc import ABCMeta
 from copy import copy
 import functools
 import types
@@ -22,47 +21,6 @@ __maintainer__ = "Jan Janssen"
 __email__ = "janssen@mpie.de"
 __status__ = "production"
 __date__ = "Sep 1, 2017"
-
-
-class Singleton(ABCMeta):
-    """
-    Implemented with suggestions from
-
-    http://stackoverflow.com/questions/6760685/creating-a-singleton-in-python
-
-    """
-
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
-
-
-def static_isinstance(obj, obj_type):
-    """
-    A static implementation of isinstance() - instead of comparing an object and a class, the object is compared to a
-    string, like 'pyiron_base.jobs.job.generic.GenericJob' or a list of strings.
-
-    Args:
-        obj: the object to check
-        obj_type (str/list): object type as string or a list of object types as string.
-
-    Returns:
-        bool: [True/False]
-    """
-    if not hasattr(obj, "__mro__"):
-        obj = obj.__class__
-    obj_class_lst = [
-        ".".join([subcls.__module__, subcls.__name__]) for subcls in obj.__mro__
-    ]
-    if isinstance(obj_type, list):
-        return any([obj_type_element in obj_class_lst for obj_type_element in obj_type])
-    elif isinstance(obj_type, str):
-        return obj_type in obj_class_lst
-    else:
-        raise TypeError()
 
 
 class Deprecator:
@@ -234,81 +192,3 @@ class Deprecator:
 
 deprecate = Deprecator()
 deprecate_soon = Deprecator(pending=True)
-
-
-class ImportAlarm:
-    """
-    In many places we have try/except loops around imports. This class is meant to accompany that code so that users
-    get an early warning when they instantiate a job that won't work when run.
-
-    Example:
-
-    >>> try:
-    ...     from mystery_package import Enigma, Puzzle, Conundrum
-    ...     import_alarm = ImportAlarm()
-    >>> except ImportError:
-    >>>     import_alarm = ImportAlarm(
-    ...         "MysteryJob relies on mystery_package, but this was unavailable. Please ensure your python environment "
-    ...         "has access to mystery_package, e.g. with `conda install -c conda-forge mystery_package`"
-    ...     )
-    ...
-    >>> class MysteryJob(GenericJob):
-    ...     @import_alarm
-    ...     def __init__(self, project, job_name)
-    ...         super().__init__()
-    ...         self.riddles = [Enigma(), Puzzle(), Conundrum()]
-
-    This class is also a context manager that can be used as a short-cut, like this:
-
-    >>> with ImportAlarm("MysteryJob relies on mystery_package, but this was unavailable.") as import_alarm:
-    ...     import mystery_package
-
-    If you do not use `import_alarm` as a decorator, but only to get a consistent warning message, call
-    :meth:`.warn_if_failed()` after the with statement.
-
-    >>> import_alarm.warn_if_failed()
-    """
-
-    def __init__(self, message=None):
-        """
-        Initialize message value.
-
-        Args:
-            message (str): What to say alongside your ImportError when the decorated function is called. (Default is
-                None, which says nothing and raises no error.)
-        """
-        self.message = message
-
-    def __call__(self, func):
-        return self.wrapper(func)
-
-    def wrapper(self, function):
-        @functools.wraps(function)
-        def decorator(*args, **kwargs):
-            self.warn_if_failed()
-            return function(*args, **kwargs)
-
-        return decorator
-
-    def warn_if_failed(self):
-        """
-        Print warning message if import has failed.  In case you are not using `ImportAlarm` as a decorator you can call
-        this method manually to trigger the warning.
-        """
-        if self.message is not None:
-            warnings.warn(self.message, category=ImportWarning)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_type == exc_value == traceback == None:
-            # import successful, so silence our warning
-            self.message = None
-            return
-        if issubclass(exc_type, ImportError):
-            # import broken; retain message, but suppress error
-            return True
-        else:
-            # unrelated error during import, re-raise
-            return False
